refactor: move release target selection
