+++ src/update/install/archive.rs
@@ -7,28 +7,6 @@ use flate2::read::GzDecoder;
 
 use super::UpdateError;
 
-pub(super) fn release_asset_name(
-    version: &str,
-    os: &str,
-    arch: &str,
-) -> Result<String, UpdateError> {
-    let target = match (os, arch) {
-        ("linux", "x86_64") => "x86_64-unknown-linux-gnu",
-        ("linux", "aarch64") => "aarch64-unknown-linux-gnu",
-        ("macos", "x86_64") => "x86_64-apple-darwin",
-        ("macos", "aarch64") => "aarch64-apple-darwin",
-        ("windows", "x86_64") => "x86_64-pc-windows-msvc",
-        _ => {
-            return Err(UpdateError::UnsupportedPlatform {
-                os: os.to_string(),
-                arch: arch.to_string(),
-            });
-        }
-    };
-    let extension = if os == "windows" { "zip" } else { "tar.gz" };
-    Ok(format!("tuicr-{version}-{target}.{extension}"))
-}
-
 pub(super) fn extract_binary(asset_name: &str, archive: &[u8]) -> Result<Vec<u8>, UpdateError> {
     match asset_name {
         name if name.ends_with(".tar.gz") => extract_tar_gz(name, archive),
