+++ src/update/install/source.rs
@@ -1,5 +1,7 @@
 use semver::Version;
 
+use super::UpdateError;
+
 pub(super) fn package_repository_url() -> &'static str {
     env!("CARGO_PKG_REPOSITORY")
         .trim_end_matches('/')
@@ -15,6 +17,28 @@ pub(super) fn release_api_url(version: Option<&Version>) -> String {
         |version| format!("tags/v{version}"),
     );
     format!("https://api.github.com/repos/{repository}/releases/{release}")
+}
+
+pub(super) fn release_asset_name(
+    version: &str,
+    os: &str,
+    arch: &str,
+) -> Result<String, UpdateError> {
+    let target = match (os, arch) {
+        ("linux", "x86_64") => "x86_64-unknown-linux-gnu",
+        ("linux", "aarch64") => "aarch64-unknown-linux-gnu",
+        ("macos", "x86_64") => "x86_64-apple-darwin",
+        ("macos", "aarch64") => "aarch64-apple-darwin",
+        ("windows", "x86_64") => "x86_64-pc-windows-msvc",
+        _ => {
+            return Err(UpdateError::UnsupportedPlatform {
+                os: os.to_string(),
+                arch: arch.to_string(),
+            });
+        }
+    };
+    let extension = if os == "windows" { "zip" } else { "tar.gz" };
+    Ok(format!("tuicr-{version}-{target}.{extension}"))
 }
 
 pub(super) fn release_asset_url(version: &str, asset_name: &str) -> String {
